Add iris.gluon shortcut, consolidate docs into host/device split
- Add iris/gluon.py re-export module so users can write
  `from iris.gluon import IrisDeviceCtx` instead of the verbose
  `from iris.device.gluon.context import IrisDeviceCtx`
- Update all tests, examples, and docs to use the short path
- Consolidate duplicate host-side doc pages (class, tensor-creation,
  CCL) into docs/reference/host/ since the Iris host class is unified
- Keep backend-specific device-side pages under triton/ and gluon/
