+++ docs/reference/api-reference.md
@@ -1,25 +1,21 @@
 # API Reference
 
-Explore Iris APIs. The reference is broken down into focused sections to mirror common workflows:
+Explore Iris APIs. The host-side API (class, tensor creation, CCL) is identical across all backends. Only the device-side API differs between Triton and Gluon.
 
-- The `Iris` class itself (constructor and helper utilities)
-- Tensor-like creation methods on the `Iris` context
-- Triton device-side functions for remote memory ops and atomics
-- Collective communication operations (CCL)
-- Fused GEMM + CCL operations
-- Experimental Gluon APIs (using `@aggregate` and `@gluon.jit`)
+## Host API (All Backends)
 
-Use the links below to navigate:
+- [Iris Class](host/class.md)
+- [Tensor Creation](host/tensor-creation.md)
+- [Collective Communication (CCL)](host/ccl.md)
 
-- [Triton](triton/overview.md)
-  - [Iris Class](triton/class.md)
-  - [Tensor Creation](triton/tensor-creation.md)
-  - [Device Functions](triton/device-functions.md)
-  - [Collective Communication (CCL)](triton/ccl.md)
-  - [Fused GEMM + CCL Operations](triton/ops.md)
-- [Gluon (Experimental)](gluon/overview.md)
-  - [Iris Class](gluon/class.md)
-  - [Tensor Creation](gluon/tensor-creation.md)
-  - [Device Functions](gluon/device-functions.md)
-  - [Collective Communication (CCL)](gluon/ccl.md)
+## Triton Backend
+
+- [Overview](triton/overview.md)
+- [Device Functions](triton/device-functions.md)
+- [Fused GEMM + CCL Operations](triton/ops.md)
+
+## Gluon Backend (Experimental)
+
+- [Overview](gluon/overview.md)
+- [Device Functions](gluon/device-functions.md)
