+++ docs/reference/host/class.md
@@ -1,15 +1,5 @@
 # Iris Class
 
-```{warning}
-The Gluon API is **experimental** and may undergo breaking changes in future releases.
-```
-
-## Requirements
-
-The Gluon backend requires:
-- **ROCm 7.0** or later
-- **Triton commit `aafec417bded34db6308f5b3d6023daefae43905`** or later
-
 ## Factory Function
 
 Prefer using the convenience factory over calling the constructor directly:
@@ -21,9 +11,9 @@ Prefer using the convenience factory over calling the constructor directly:
 ## Core Methods
 
 ```{eval-rst}
+.. automethod:: iris.host.iris.Iris.get_heap_bases
 .. automethod:: iris.host.iris.Iris.get_device_context
 .. automethod:: iris.host.iris.Iris.get_backend
-.. automethod:: iris.host.iris.Iris.get_heap_bases
 .. automethod:: iris.host.iris.Iris.barrier
 .. automethod:: iris.host.iris.Iris.get_device
 .. automethod:: iris.host.iris.Iris.get_cu_count
@@ -36,10 +26,17 @@ Prefer using the convenience factory over calling the constructor directly:
 Use Iris-aware logging that automatically annotates each message with the current rank and world size. This is helpful when debugging multi-rank programs.
 
 ```{eval-rst}
+.. autofunction:: iris.host.logging.logging.set_logger_level
 .. automethod:: iris.host.iris.Iris.debug
 .. automethod:: iris.host.iris.Iris.info
 .. automethod:: iris.host.iris.Iris.warning
 .. automethod:: iris.host.iris.Iris.error
+```
+
+## Utility Functions
+
+```{eval-rst}
+.. autofunction:: iris.host.platform.utils.do_bench
 ```
 
 ## Broadcast Helper
@@ -49,6 +46,3 @@ Broadcast data from a source rank to all ranks. This method automatically detect
 ```{eval-rst}
 .. automethod:: iris.host.iris.Iris.broadcast
 ```
-
-
-
