+++ docs/reference/host/tensor-creation.md
@@ -1,9 +1,5 @@
 # Tensor Creation
 
-```{warning}
-The Gluon API is **experimental** and may undergo breaking changes in future releases.
-```
-
 APIs on `Iris` that create and initialize tensors on the Iris symmetric heap.
 
 ```{eval-rst}
@@ -11,11 +7,18 @@ APIs on `Iris` that create and initialize tensors on the Iris symmetric heap.
 .. automethod:: iris.host.iris.Iris.zeros_like
 .. automethod:: iris.host.iris.Iris.ones
 .. automethod:: iris.host.iris.Iris.full
+.. automethod:: iris.host.iris.Iris.empty
+.. automethod:: iris.host.iris.Iris.rand
+.. automethod:: iris.host.iris.Iris.randn
+.. automethod:: iris.host.iris.Iris.uniform
+.. automethod:: iris.host.iris.Iris.randint
+.. automethod:: iris.host.iris.Iris.linspace
+.. automethod:: iris.host.iris.Iris.arange
 ```
 
 ## Symmetric Heap Utilities
 
 ```{eval-rst}
+.. automethod:: iris.host.iris.Iris.as_symmetric
 .. automethod:: iris.host.iris.Iris.is_symmetric
 ```
-
