--- docs/reference/triton/ccl.md
@@ -1,28 +1,0 @@
-# Collective Communication Operations
-
-Collective communication operations accessible via the `ccl` attribute on the `Iris` instance (e.g. `ctx.ccl.all_reduce(...)`).
-
-## all_to_all
-```{eval-rst}
-.. automethod:: iris.iris.Iris.CCL.all_to_all
-```
-
-## all_gather
-```{eval-rst}
-.. automethod:: iris.iris.Iris.CCL.all_gather
-```
-
-## all_reduce_preamble
-```{eval-rst}
-.. automethod:: iris.iris.Iris.CCL.all_reduce_preamble
-```
-
-## all_reduce
-```{eval-rst}
-.. automethod:: iris.iris.Iris.CCL.all_reduce
-```
-
-## reduce_scatter
-```{eval-rst}
-.. automethod:: iris.iris.Iris.CCL.reduce_scatter
-```
