--- docs/reference/triton/tensor-creation.md
@@ -1,24 +1,0 @@
-# Tensor Creation
-
-APIs on `Iris` that create and initialize tensors on the Iris symmetric heap.
-
-```{eval-rst}
-.. automethod:: iris.iris.Iris.zeros
-.. automethod:: iris.iris.Iris.zeros_like
-.. automethod:: iris.iris.Iris.ones
-.. automethod:: iris.iris.Iris.full
-.. automethod:: iris.iris.Iris.empty
-.. automethod:: iris.iris.Iris.rand
-.. automethod:: iris.iris.Iris.randn
-.. automethod:: iris.iris.Iris.uniform
-.. automethod:: iris.iris.Iris.randint
-.. automethod:: iris.iris.Iris.linspace
-.. automethod:: iris.iris.Iris.arange
-```
-
-## Symmetric Heap Utilities
-
-```{eval-rst}
-.. automethod:: iris.iris.Iris.as_symmetric
-.. automethod:: iris.iris.Iris.is_symmetric
-```
