+++ examples/32_gluon_all_gather_tracing/all_gather_tracing.py
@@ -43,7 +43,7 @@ from triton.experimental import gluon
 from triton.experimental.gluon import language as gl
 
 import iris
-from iris.device.gluon.context import IrisDeviceCtx
+from iris.gluon import IrisDeviceCtx
 from iris.host.tracing.events import TraceEvent
