+++ iris/experimental/__init__.py
@@ -12,7 +12,7 @@ Current experimental features:
 
 Usage:
     >>> import iris
-    >>> from iris.device.gluon.context import IrisDeviceCtx
+    >>> from iris.gluon import IrisDeviceCtx
     >>> from triton.experimental import gluon
     >>> from triton.experimental.gluon import language as gl
     >>>
