+++ tests/unittests/test_load_gluon.py
@@ -6,7 +6,7 @@ import pytest
 from triton.experimental import gluon
 from triton.experimental.gluon import language as gl
 import iris
-from iris.device.gluon.context import IrisDeviceCtx
+from iris.gluon import IrisDeviceCtx
 
 
 @gluon.jit
